+++ app/app.component.html
@@ -1,0 +1,6 @@
+<h1>{{title}}</h1>
+   <nav>
+     <a routerLink="/dashboard" routerLinkActive="active">Dashboard</a>
+     <a routerLink="/heroes" routerLinkActive="active">Heroes</a>
+   </nav>
+   <router-outlet></router-outlet>
