+++ app/app.module.ts
@@ -1,11 +1,10 @@
 import { NgModule }      from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule }   from '@angular/forms';
-import { RouterModule }   from '@angular/router';
 
 import { AppComponent }   from './app.component';
 import { HeroDetailComponent } from './modules/hero-detail/hero-detail.component';
-import { HeroesComponent } from './modules/heroes.component';
+import { HeroesComponent } from './modules/heroes/heroes.component';
 import { DashboardComponent } from './modules/dashboard/dashboard.component';
 import { HeroService } from './services/hero.service';
