+++ app/modules/heroes/heroes.component.ts
@@ -1,0 +1,36 @@
+import { Component } from '@angular/core';
+import { OnInit } from '@angular/core';
+import { Router } from '@angular/router';
+
+import { Hero } from '../../models/hero';
+
+import { HeroDetailComponent } from '../hero-detail/hero-detail.component';
+import { HeroService } from '../../services/hero.service';
+
+@Component({
+  moduleId: module.id,
+  selector: 'my-heroes',
+  templateUrl:'heroes.component.html',
+  styleUrls: ['heroes.component.css']
+})
+
+export class HeroesComponent implements OnInit{
+  selectedHero : Hero;
+  heroes : Hero[];
+
+  constructor(
+    private router: Router,
+    private heroService: HeroService) { }
+
+  ngOnInit(): void {
+    this.heroService.getHeroesSlowly().then(heroes => this.heroes = heroes);
+  }
+
+  onSelect(hero: Hero): void {
+    this.selectedHero = hero;
+  }
+
+  gotoDetail(): void {
+    this.router.navigate(['/detail', this.selectedHero.id]);
+  }
+}
