+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "angular-quickstart",
+  "name": "angular-super-hero-tour",
   "version": "1.0.0",
   "scripts": {
     "start": "tsc && concurrently \"tsc -w\" \"lite-server\" ",
