+++ styles.css
@@ -12,3 +12,12 @@ h2, h3 {
 body {
   margin: 2em;
 }
+body, input[text], button {
+  color: #888;
+  font-family: Cambria, Georgia;
+}
+/* . . . */
+/* everywhere else */
+* {
+  font-family: Arial, Helvetica, sans-serif;
+}
